refactor: extract theme detection to utils module
