+++ src/tray.rs
@@ -4,7 +4,6 @@
 //! Linux uses the ksni-based implementation in tray_linux.rs.
 
 use crossbeam_channel::Receiver;
-use image::RgbaImage;
 use std::sync::{Arc, RwLock};
 use tray_icon::{
     menu::{CheckMenuItem, Menu, MenuEvent, MenuItem, PredefinedMenuItem},
@@ -13,36 +12,7 @@ use tray_icon::{
 
 use crate::autostart;
 use crate::state::AppState;
-
-/// Detects if the system is using light mode.
-fn is_light_mode() -> bool {
-    match dark_light::detect() {
-        Ok(dark_light::Mode::Light) => true,
-        Ok(dark_light::Mode::Unspecified) => {
-            // Default to dark mode (white icon) when unspecified
-            false
-        }
-        Ok(dark_light::Mode::Dark) => false,
-        Err(_) => {
-            // On error, default to dark mode (white icon)
-            false
-        }
-    }
-}
-
-/// Creates an inverted (dark) version of the icon for light mode.
-/// Preserves alpha channel while inverting RGB values.
-fn create_dark_icon(image: &RgbaImage) -> RgbaImage {
-    let mut dark = image.clone();
-    for pixel in dark.pixels_mut() {
-        // Invert RGB, keep alpha
-        pixel[0] = 255 - pixel[0]; // R
-        pixel[1] = 255 - pixel[1]; // G
-        pixel[2] = 255 - pixel[2]; // B
-                                   // pixel[3] = alpha, keep as-is
-    }
-    dark
-}
+use crate::utils::{create_dark_icon, is_light_mode};
 
 pub enum TrayCommand {
     Quit,
