+++ src/utils.rs
@@ -674,3 +674,33 @@ impl MediaType {
         }
     }
 }
+
+/// Detects if the system is using light mode.
+pub fn is_light_mode() -> bool {
+    match dark_light::detect() {
+        Ok(dark_light::Mode::Light) => true,
+        Ok(dark_light::Mode::Unspecified) => {
+            // Default to dark mode (white icon) when unspecified
+            false
+        }
+        Ok(dark_light::Mode::Dark) => false,
+        Err(_) => {
+            // On error, default to dark mode (white icon)
+            false
+        }
+    }
+}
+
+/// Creates an inverted (dark) version of the icon for light mode.
+/// Preserves alpha channel while inverting RGB values.
+pub fn create_dark_icon(image: &image::RgbaImage) -> image::RgbaImage {
+    let mut dark = image.clone();
+    for pixel in dark.pixels_mut() {
+        // Invert RGB, keep alpha
+        pixel[0] = 255 - pixel[0]; // R
+        pixel[1] = 255 - pixel[1]; // G
+        pixel[2] = 255 - pixel[2]; // B
+                                   // pixel[3] = alpha, keep as-is
+    }
+    dark
+}
